Remove column alpha-ordering
A check is performed in `flattening.flatten_schema` to ensure that there are no duplicate columns by name. A sorted dict called `sorted_items` is created to perform the check, however the flattened schema does not need to be sorted, so the unsorted version (`items`) should be returned by this function.